Rename to GalleryEditorRenderer
